Move validation to types package
